Separate TimeLossField, TelescopesStates, JiraOBSTicketsTable, CSCStates components from CreateNightReport for reusability.
